factor out NaiveDatetime
