+++ tests/db/test_sqlmodel_models.py
@@ -1,10 +1,10 @@
 from datetime import datetime, timedelta, timezone
 
+from luthien_control.db.naive_datetime import NaiveDatetime
 from luthien_control.db.sqlmodel_models import (
     ClientApiKey,
     ControlPolicy,
     LuthienLog,
-    NaiveDatetime,
 )
 
 
@@ -165,7 +165,7 @@ def test_client_api_key_creation_all_fields():
 
 def test_luthien_log_creation_required_fields():
     """Test successful creation of LuthienLog with required fields."""
-    now_utc = NaiveDatetime.now()  # Use NaiveDatetime
+    now_utc = NaiveDatetime.now()
     log = LuthienLog(transaction_id="tx-abc-123", datatype="test_event")
 
     assert log.id is None
@@ -180,7 +180,7 @@ def test_luthien_log_creation_required_fields():
 
 def test_luthien_log_creation_all_fields():
     """Test successful creation of LuthienLog with all fields specified."""
-    log_time = NaiveDatetime(2023, 6, 15, 14, 30, 0)  # Use NaiveDatetime
+    log_time = NaiveDatetime(2023, 6, 15, 14, 30, 0)
     log_data_content = {"event_details": "something happened"}
     log_notes_content = {"source": "test_suite"}
 
@@ -203,48 +203,11 @@ def test_luthien_log_creation_all_fields():
 
 def test_luthien_log_repr():
     """Test the __repr__ method of LuthienLog."""
-    log_time = NaiveDatetime(2023, 1, 1, 0, 0, 0)  # Use NaiveDatetime
+    log_time = NaiveDatetime(2023, 1, 1, 0, 0, 0)
     log = LuthienLog(id=1, transaction_id="repr-tx", datetime=log_time, datatype="repr_test")
     # No timezone in repr
     expected_repr = "<LuthienLog(id=1, transaction_id='repr-tx', datetime='2023-01-01 00:00:00', datatype='repr_test')>"
     assert repr(log) == expected_repr
-
-
-def test_naive_datetime_strips_timezone():
-    """Test that NaiveDatetime automatically strips timezone info."""
-    # Test with timezone-aware datetime
-    aware_dt = datetime(2023, 1, 1, 12, 0, 0, tzinfo=timezone.utc)
-    naive_dt = NaiveDatetime(aware_dt)
-
-    assert naive_dt.tzinfo is None
-    assert naive_dt.year == 2023
-    assert naive_dt.month == 1
-    assert naive_dt.day == 1
-    assert naive_dt.hour == 12
-
-
-def test_naive_datetime_preserves_naive():
-    """Test that NaiveDatetime preserves already-naive datetimes."""
-    # Test with already-naive datetime
-    original_naive = datetime(2023, 1, 1, 12, 0, 0)
-    naive_dt = NaiveDatetime(original_naive)
-
-    assert naive_dt.tzinfo is None
-    assert naive_dt == original_naive
-
-
-def test_naive_datetime_converts_non_utc_timezone():
-    """Test that NaiveDatetime converts non-UTC timezones to naive UTC."""
-    from datetime import timedelta
-
-    # Create a timezone-aware datetime in EST (UTC-5)
-    est = timezone(timedelta(hours=-5))
-    est_dt = datetime(2023, 1, 1, 12, 0, 0, tzinfo=est)  # 12 PM EST = 5 PM UTC
-
-    naive_dt = NaiveDatetime(est_dt)
-
-    assert naive_dt.tzinfo is None
-    assert naive_dt.hour == 17  # Should be converted to 5 PM UTC
 
 
 def test_luthien_log_with_timezone_aware_datetime():
@@ -264,60 +227,3 @@ def test_luthien_log_with_timezone_aware_datetime():
     assert log.datetime.month == 6
     assert log.datetime.day == 15
     assert log.datetime.hour == 14  # Should preserve the UTC time
-
-
-def test_naive_datetime_normal_constructor():
-    """Test that NaiveDatetime works with normal datetime constructor arguments."""
-    naive_dt = NaiveDatetime(2023, 1, 1, 12, 0, 0)
-
-    assert naive_dt.tzinfo is None
-    assert naive_dt.year == 2023
-    assert naive_dt.month == 1
-    assert naive_dt.day == 1
-    assert naive_dt.hour == 12
-
-
-def test_tx_logging_policy_datetime_handling():
-    """Test that demonstrates our solution prevents the original timezone issue."""
-    from datetime import datetime, timezone
-
-    from luthien_control.db.sqlmodel_models import NaiveDatetime
-
-    # Simulate what TxLoggingPolicy does now - creates timezone-aware datetime
-    # but converts it to naive before passing to LuthienLog
-    timezone_aware_dt = datetime.now(timezone.utc)
-
-    # Our fix: use NaiveDatetime constructor to strip timezone
-    naive_dt = NaiveDatetime(timezone_aware_dt)
-
-    # Verify it's properly converted
-    assert naive_dt.tzinfo is None
-    assert isinstance(naive_dt, NaiveDatetime)
-
-    # This naive datetime can now be safely used in LuthienLog
-    log = LuthienLog(
-        transaction_id="tx-policy-test",
-        datatype="policy_test_event",
-        datetime=naive_dt,
-    )
-
-    # The datetime should be naive and safe for database insertion
-    assert log.datetime.tzinfo is None
-    assert log.datetime.year == timezone_aware_dt.year
-    assert log.datetime.month == timezone_aware_dt.month
-    assert log.datetime.day == timezone_aware_dt.day
-
-
-def test_naive_datetime_now():
-    """Test that NaiveDatetime.now() returns current naive UTC time."""
-
-    before = NaiveDatetime.now()
-    naive_now = NaiveDatetime.now()
-    after = NaiveDatetime.now()
-
-    # Should be naive
-    assert naive_now.tzinfo is None
-
-    # Should be close to current time (within 1 second)
-    assert before <= naive_now <= after or abs((naive_now - before).total_seconds()) < 1
-    assert isinstance(naive_now, NaiveDatetime)
